# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehouse Management API",
        "description": "API quản lý kho dùng Laravel 12, MySQL, Sanctum, Spatie Permission và Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập",
                "operationId": "1eaad5044f49331e2e4a269354b56d7b",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng xuất",
                "operationId": "f429f77c76ed94c688e77779a9621005",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Thông tin tài khoản hiện tại",
                "operationId": "21900d7507dd7fa34d2eca27483f1c17",
                "responses": {
                    "200": {
                        "description": "Thông tin người dùng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/warehouses": {
            "get": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Danh sách kho và tìm theo địa chỉ",
                "operationId": "27a29ac123100e7ea154cdb31e484722",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách kho"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Tạo kho",
                "operationId": "2ec0b1fe94eeed133631c0c3d1855c35",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Tạo kho thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/warehouses/{warehouse}": {
            "get": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Chi tiết kho",
                "operationId": "69caad8bddbe566e28f90b7d494f4457",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết kho"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Cập nhật kho",
                "operationId": "a6e9932c78a279fbcbf67e7f78354b02",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Xóa kho",
                "operationId": "148a70a3c97853a9819216fd6f346bdd",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Nhập token theo dạng: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}